Move all application components to a dedicated 'components' directory
